Add Roster
Roster was a function in planning that would take all the people signing up as "new" (freshman) and put them into a spreadsheet for one of the other officers to use as an easy import instead of manually entering everybody's handwritten names.
Unfortunately, this never caught on. But I thought it would have been a good idea.